Consolidate config_manager and config_loader into singular configuration module
